power:qpnp-charger: extended charge time
from 30s to 100s when battery status is near to full

When charger is online for a long time, such as 12 hour, the battery status changes from charging to full and resumes charging.
The reason is that battery is not full at the first full state. So increse charge time from 30s to 100s when battery status is
near to full. This methods will reduce the number of recharging times

BUG:24316771
Change-Id: If4a98f2b07502d184d74d04acb3db443c56e64ab
Signed-off-by: l00228880 <l00228880@notesmail.huawei.com>